Add templates the default option
Add a button to make one of the templates the default option that
displayed when one of the filter criteria isn't meet.

Add a check box say if the template is for male female or both and have
that info display in the dashoboard area.

Have a drop down menu added for state selection.
Have them type in the city name.
Have this show on the dashboard area also.

If the ip address of the user logging in is for that state have it
display that page.

DB Change:
ALTER TABLE `templates` ADD `is_default` BOOLEAN NOT NULL AFTER
`widgets`, ADD `is_male` BOOLEAN NOT NULL AFTER `is_default`, ADD
`is_female` BOOLEAN NOT NULL AFTER `is_male`, ADD `is_both` BOOLEAN NOT
NULL AFTER `is_female`, ADD `state` VARCHAR(30) NOT NULL AFTER
`is_both`, ADD `city` VARCHAR(30) NOT NULL AFTER `state`;

Revert "Add templates the default option"

This reverts commit
f044f77bb5431d700649ad08b7908a1176929457.

Add templates the default option

Add a button to make one of the
templates the default option that
displayed when one of the filter
criteria isn't meet.

Add a check box say if the template is for male
female or both and have
that info display in the dashoboard area.

Have
a drop down menu added for state selection.
Have them type in the city
name.
Have this show on the dashboard area also.

If the ip address of
the user logging in is for that state have it
display that page.